Refactor: folder structure
