+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import RootNavigation from "./routes/rootNavigation.jsx";
+import RootNavigation from "./routes/RootNavigation.jsx";
 
 
 export default function App() {
