+++ src/pages/body/Body.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
 
-import Aside from "../components/aside";
+import Aside from "../../components/ui/Aside.jsx";
 
 import "./body.css";
